+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Extend list functionality:
   * methods: `asMapWithKeys`
 * Extend map functionality:
-  * methods: `filter`, `filterKeys`, `filterValues`, `where`, `whereKey`, `whereValue`
+  * methods: `filter`, `filterKeys`, `filterValues`, `single`, `singleKey`, `singleValue`, `where`, `whereKey`, `whereValue`
 
 ## 1.1.0